+++ CHANGES.md
@@ -11,6 +11,7 @@ v3.0.3 - YYYY-MM-DD
 - Fixed HTTP state for POST/PUT with an empty message body (Issue #149)
 - Fixed the example RPM spec file (Issue #150)
 - Fixed potential buffer overflow in `cupsCopyDestConflicts`.
+- Added new PWG media sizes.
 
 
 v3.0.2 - 2026-06-05